+++ tests/IntegratedTest.php
@@ -11,6 +11,7 @@ class IntegratedTest extends TestCase
         Artisan::call('packager:new', ['vendor' => 'MyVendor', 'name' => 'MyPackage']);
 
         $this->seeInConsoleOutput('Package created successfully!');
+        $this->assertTrue(is_link(base_path('vendor/myvendor/mypackage')));
     }
 
     public function test_new_package_is_installed()
@@ -20,6 +21,15 @@ class IntegratedTest extends TestCase
         $composer = file_get_contents(base_path('composer.json'));
 
         $this->assertStringContainsString('MyVendor/MyPackage', $composer);
+    }
+
+    public function test_new_package_studly_install()
+    {
+        Artisan::call('packager:new', ['vendor' => 'my-vendor', 'name' => 'my-package']);
+
+        $this->seeInConsoleOutput('Package created successfully!');
+        $this->assertStringContainsString('my-vendor/my-package', file_get_contents(base_path('composer.json')));
+        $this->assertTrue(is_file(base_path('packages/my-vendor/my-package/src/MyPackageServiceProvider.php')));
     }
 
     public function test_new_package_is_installed_from_custom_skeleton()
@@ -35,12 +45,30 @@ class IntegratedTest extends TestCase
         $this->assertStringContainsString('AnotherVendor/AnotherPackage', $composer);
     }
 
-    public function test_get_existing_package()
+    public function test_get_package()
     {
         Artisan::call('packager:get',
             ['url' => 'https://github.com/Jeroen-G/packager-skeleton', 'vendor' => 'MyVendor', 'name' => 'MyPackage']);
 
         $this->seeInConsoleOutput('Package downloaded successfully!');
+    }
+
+    public function test_get_existing_package_with_git()
+    {
+        Artisan::call('packager:git',
+            ['url' => 'https://github.com/Seldaek/monolog', 'vendor' => 'monolog', 'name' => 'monolog']);
+
+        $this->seeInConsoleOutput('Package cloned successfully!');
+        $this->assertTrue(is_link(base_path('vendor/monolog/monolog')));
+    }
+
+    public function test_get_existing_package_with_get()
+    {
+        Artisan::call('packager:get',
+            ['url' => 'https://github.com/Seldaek/monolog', 'vendor' => 'monolog', 'name' => 'monolog']);
+
+        $this->seeInConsoleOutput('Package downloaded successfully!');
+        $this->assertTrue(is_link(base_path('vendor/monolog/monolog')));
     }
 
     public function test_list_packages()